narratives: Convert Markdown → HTML before splitting into separate slides
Fixes handling of intra-document references like links and images that
reference a definition found elsewhere in the file, e.g. before the
first slide or on a different slide than the link/image reference.  Such
features are soon to be used for embedding images (see related PR
below).

Anyone previewing a document with such features outside of Auspice's
narrative rendering would see the link/image references work, but they'd
be broken when viewed thru Auspice.  Since narrative documents are
intended to gracefully degrade with conventional Markdown rendering and
still be a viewable document (e.g. on GitHub), it's nice to address this
issue for that reason alone.

It's also nice to use the DOM as a structured way to split apart the
document into slides, rather than relying on string manipulation and
parsing.  An alternative approach would be to use a Markdown parser that
emits an AST or stream of nodes as an intermediate we could split into
slides while still preserving references/definitions, but that seemed
like a bigger change (and I'd have to go survey the parsers available
instead of using the browser's built-in DOM).

While the entire internals of parseNarrativeBody() change, the shape of
the output "slide objects" is the same.  The one minor exception is that
slides will always have a "mainDisplayMarkdown" property (though it will
usually be the empty string) even if they don't have a corresponding
code block in the source text.  All downstream usages I found condition
not on property existence but on truthiness of the value, so this should
result in no functional change.

Resolves: <https://github.com/nextstrain/auspice/issues/1599>
Related-to: <https://github.com/nextstrain/cli/pull/235>