WIP - save animation frames as PNG
We've long wanted the ability to save animations for later use (in presentations etc)
I spent an hour trying to implement this by saving each frame as a PNG using
html2canvas with the intention of then using ffmpeg to stitch & crop them
into a video. It kind-of worked... for small phylogenies the images look good
but for larger ones the tree panel was blank (apart from the legend). I presume
this is a DOM-complexity thing, but I haven't looked into it!