Use newer version of CSSTransitionGroup
Now that NPM 7 tries to install peer dependencies, having an old version
of react-addons-css-transition-group alongside a newer version of React
will not work with `npm install` as it used to.

However, simply upgrading the old package does not work since there are
no newer versions compatible with React 16. Instead, it has been
deprecated in favor of react-transition-group¹. Luckily, version 1 of
the new package provides a drop-in replacement.

Commands used²:

    npm uninstall react-addons-css-transition-group --legacy-peer-deps
    npm install react-transition-group@1 --legacy-peer-deps

¹ https://www.npmjs.com/package/react-addons-css-transition-group
² --legacy-peer-deps is necessary with NPM 7 until other peer dependency
  issues are resolved by future commits.