narratives: Remove legacy support for server-side parsing from the built-in server
Removes support for requests of the form:

    /charon/getNarrative?prefix=…&type=json

Leaving only support for requests of the form:

    /charon/getNarrative?prefix=…&type=md

Although Auspice switched to client-side parsing in 2.18.0 (5b941343,
Aug 2020), the server-side parsing support was preserved in the name of
backwards compatibility.  That would have allowed a newer Auspice server
to serve narratives to an older Auspice client, though I'm not sure if
that was ever done in practice outside of development and testing.

Removing this legacy support allows us to directly use client-side
features in src/util/parseNarrative.js that aren't available server-side
(such as the built-in HTML → DOM parsing of browsers), since that code
is no longer used in both contexts.