Remove prepare script
The prepare script is automatically run with every call to npm ci / npm
install, among other scripts¹. As it is just an alias to the build
script (which can take several minutes), it should be safe to remove
this automation and speed up the dev experience.

Some parts of the README and CI workflow had relied on the prepare
script automation. Updated accordingly.

¹ https://docs.npmjs.com/cli/v8/using-npm/scripts?v=true#life-cycle-operation-order