Add 'none' option to hide tip labels
Adds a new 'none' option to turn off the tip labels.

We should consider using a Symbol for the 'none' value so that it can't
clash with potential coloring values, but then it's not as
straightforward to specify this option via the URL query. This does mean
any coloring with key 'none' cannot be used as a tip label.