docs: Restore literal tabs to the TSV examples in the rST source
These were inadvertently lost in "Convert markdown pages to
reStructuredText" (7b8e45e2) which converted them to spaces.

Note that these tabs still get expanded to spaces when rendering the rST
to HTML, and thus copying the example from the HTML page to paste into a
file will not produce a TSV.  I'll address this in a subsequent commit.