docs: Render example TSV files as tables and provide download links
This makes the TSV examples more readable on the page (i.e. the columns
line up unlike the raw TSV) while still retaining the ability for a
reader to get a copy of the example file to tinker with.

We could use the tab-width option of the literalinclude directive to
line up the columns, but that would break the ability to copy/paste the
contents into a new TSV file, as the tabs get expanded to spaces when
tab-width is used.