Use npm command to apply version bump to NPM files
Previously, a custom perl script was used to apply the changes
package.json. However, changes were not applied to package-lock.json.

Using the same perl script on package-lock.json wouldn't be right
because the "version" string can potentially match versions of other
packages. Another custom script is possible, but instead we can use a
npm command designed for this type of update.