Fix bounds matching for zero values
This fixes a common bug for continuous scales where nodes have (valid)
zero values. The bound for the first (smallest) legend entry would have
zero as it's lower value, and as we strictly matched `v \in [a,b)`
the node was never included as a match. This presented in two places:
1. Hovering over a legend swatch would not highlight (increase the
radius of) matching nodes whose value was zero
2. The frequencies stream graph would map zeros to "unassigned" rather
than their correct legend bin.

There were a few code changes needed to case (2) as we also had
falsey checks (and 0 is falsey).

Note that the algorithm to create legend bounds is problematic and
should be rewritten.

This bug was first identified by Hugh Haddox, but others may have
noticed it over the years?!?

The /flu/seasonal/h3n2/ha/2y dataset is useful to test this bug:
* `c=ne_star` - most of the tree has zeros, none are unassigned
* `c=projected_frequency` - both unassigned and zero tips.
* `c=fitness` failure to match lowest legend entry due to the bug in
the legend bounds creation