Remove bundlesize from list of devDependencies
bundlesize has some indirect dependencies which require extra build
time. Currently, it is only used for CI, so the extra time required to
build/install is unnecessary for most use cases.

Instead of installing bundlesize alongside Auspice, install it through
npx where it is used. Installing through npx installs to a folder in the
npm cache.