download/helperFunctions: create TSV with Papa Parse
Adds a new function `createTsvString` that uses `Papa.unparse` to
convert an array of data objects to a properly formatted TSV string.
This includes an optional `columns` parameter that allows us to specify
the specific keys and their order to use as columns. I've hard-coded
the other config options to ensure that all downloaded TSVs have the
same format, but we may want to relax this in the future and add other
config parameters. I opted to use the array of objects data model for
this function to avoid bugs that can come up with mismatched orders of
header fields vs data values.

Updates the existing *TSV functions to use this new function instead of
naively joining string values with tabs. This prevents the downloaded
TSV files from being improperly formatted such as the unescaped double
quote described in #1648.