Type cast continuous values in tree
Conceptually these values only make sense if they are numeric.
Previously some implicit casting was happening, but this resulted in
bugs such as negative values not being cast and thus not displayed in
the scatterplot.

Note one slight side-effect which may be considered a regression:
previously string values which were not numbers (e.g. "abc" but not
"1.0") would show up in the hover/click info-boxes. Now such values will
be changed to `undefined` and therefore not shown in those info-boxes.

Timing of the changes introduced here (AppleM1, Firefox 111):
* nextclade/sars-cov-2/21L   7-9ms
* ncov/gisaid/global/6m      4-10ms
* zika                       2ms

Closes #1626