Fix legend matching bounds
Previously the creation of legend bounds for continuous scales used
hardcoded minimum (0) and maximum (10,000). This may result in the
first and/or last legend entry having associated bounds which could
never match any values.

Given the various locations this function is called it's easier to
set the min/max as (-)Infinity rather than consider the observed min/max.

Closes #1644
Closes #1654