wip! blindly upgrade a bunch of things with problematic peer deps on react
All of these had peer deps on react that evaluated to <18.  Upgrades for
most were available, often major version bumps so they may come with
breaking changes.  A couple things were not available and we need to
switch to alternatives.  Our calling code for libraries has not been
updated, so it's unlikely to compile or run.  But npm is now happy!