actions/measurements: add `parseMeasurementsJSON`
During loading of the measurements JSON, we do some data transformations
for easier data manipulations on the client side such as transforming
Arrays to Maps. However, this did not account for the loading of
measurements JSONs in narratives, which uses cached JSONs from Redux.
These data transformations inadvertently updated the cached JSONs in
Redux, leading to errors when we tried to load the same JSON again when
navigating back to a previous slide with a different dataset.

This commit pulls out the parsing of the measurements JSON into a new
function that is chained to `fetchJSON` so that we store the
post-transformed value of the JSON in the Redux cache. Subsequent
loading of the same JSON will no longer perform any data transformations.

Note that this will break the loading of the measurements sidecar JSON
in auspice.us since we are changing the JSON cache. Will need to update
auspice.us to use the new `parseMeasurementsJSON` function in the file
handling.