eslint: Add extension to config file
The extension-less configuration file name was deprecated more than 5
years ago¹, and the deprecation message is no longer present in the
official documentation.

Even though the extension-less file still works fine, the rename serves
to keep up-to-date with documented formats and make it clear to some
text editors that YAML is the format used in this file.

¹ https://dev.to/ohbarye/eslintrc-without-file-extension-is-deprecated-3ngg