Clarify existing scope of ESLint
Currently, the ESLint configuration file is top-level but only files
under src/ are linted according to the npm script. This is a bit
disjoint.

Two options:

1. Move all usage of ESLint to src/.
2. Work towards applying ESLint to all files in this project.

I'm going with (2) since there are some unlinted files that could
benefit from linting. The first step, done by this commit, is to update
the npm script and specify the unlinted files explicitly in
.eslintignore (also removing unused ESLint exceptions from these files).
Future commits will continue in this direction.