Tip label is used for on-hover+click info
This generally better reflects the intention of changing the tip label
key, which was previously only reflected in text drawn to the RHS of
tree nodes at certain zoom levels.

The motivating use case is the proliferation of datasets using an
accession for `node.name` but who wish to see node names of another
piece of metadata. This is possible by adding the desired name as a
node_attr and (optionally) setting this as the default tip label.

Note that the order of items in the node-clicked modal has been
slightly altered and is now more consistent across internal branches,
terminal branches & tips