Tip label is used for on-hover+click info
This generally better reflects the intention of changing the tip label
key, which was previously only reflected in text drawn to the RHS of
tree nodes at certain zoom levels.

The motivating use case is the proliferation of datasets using an
accession for `node.name` but who wish to see node names of another
piece of metadata. This is possible by adding the desired name as a
node_attr, making this a coloring, and setting the default tip label,
e.g.:

```
node_attr.strain = {value: 'Strain Name'}
display_default.tip_label = 'strain'
```