Update puppeteer and jest-puppeteer
As recommended by `npm audit`

The removal of the line in `puppeteer.setup.js` was motivated by the
introduced error "TypeError: page._client.send is not a function". I'm
reasonably confident that line was related to integration tests
removed in a previous commit.

Note that this version of puppeteer uses nullish coalescing assignment
which is not supported by node.js v14.