wip! Convert genome annotations to new format
Currently the JSON annotation block is overly simple, and there are
many features we want to parse & display in Auspice. This will
inevitably involve an update to the JSON schema, but here we introduce
a new hierarchical structure `genomeMap` to represent
Genome → Chromosome[] → Gene[] → CDS[] → CDS_Segment[]

Parsing multiple CDS segments from a single CDS in the current JSON
schema is only possible for CDSs which wrap around the origin.

We provide a default color per gene, but allow per-CDS overrides from
the JSON.

Future commits will remove the `entropy.annotations` and `entropy.geneMap`
as the intention is that `entropy.genomeMap` is the canonical source of
this information.

Note - previously auspice treated CDSs & genes as the same thing.
They are now distinguished from each other, but this change will
take time to propagate through the codebase.

Note - Typescript errors in existing code are addressed largely using "any"
types, as the plan is to remove this code as part of this PR.