[entropy] refactor SVG groups and position offsets
This is in preparation for upcoming changes to how annotations are
drawn.

The offsets (position calculations) remain the same, but now use named
constants to better convey the intention. Different elements are now
separated out so that we may move them independently -- e.g. the nav
annotations, brush and axis are no longer tied together by the same
offset.

There still remain some hardcoded pixel values, especially in the
brush code, which we will improve when we update that later on.

Wireframes are added to the brush elements to help with development.
They will be removed in an upcoming commit (on this branch), as the
zoom / brush will be significantly changed.