[entropy] Refactor tooltip
Previously we only had one tooltip (active while hovering on bars),
generated by `<InfoPanel>`. Here we abstract out the logic for
tooltip positioning & display and then delegate the contents of the
tooltip to the triggering event. This is in preparation for future
commits which will add tooltips for other events in the entropy panel.