WIP! [entropy] refactor redux state
TODO - remove copious console.log statements

Previously the entropy panel could toggle between "aa" and "nuc", and
we stored this in `state.controls.mutType`. The aim of this PR is to
be able to select a CDS (or select "nucleotide view") within the entropy
panel rather than a aa-nuc toggle. We could simply add another bit of
state to represent this, but here I've chosen to remove mutType and
introduce `state.entropy.selectedCds`. Note that `mutType` still appears
in some functions, but it is derived from `selectedCds` rather than taken
from redux state.

This involved a lot of changes to the codebase, but it is significantly
improved because of it. The Entropy panel now no subscribes to redux
colorBy state, which addresses a series of bugs related to state-change-
ordering (see removed comment from `./src/actions/colors.js`).

There should be minimal changes to the UI, but there are a few:

* The AA-NT toggle has gone from the entropy panel. AA entropy can still
be shown by using an appropriate genotype colorBy (via UI or URL).

* When switching to nuc view in the entropy panel from an aa-genotype-
colorBy, or vice vera,  we previously would change the colorBy to the
default (often "country"). This is no longer the case, as I don't think
it's that user-friendly. But it's easy to add back if people like it!

* The way the entropy panel automatically zooms upon genotype changes
isn't working quite as before. This will be a big focus of a future commit,
and the current implementation is rather ad-hoc, so I'm not paying it
too much attention here.