download: use the `node.shell.inView` property
The `node.inView` property is a temporary hold for the node's inView
property to be able to zoom in on specific branches before the
D3 PhyloTree has been constructed.

Once the D3 PhyloTree has been constructed, the inView property is
available through `node.shell.inView`. This is the property that gets
updated by `applyInViewNodesToTree` whenever the user interacts with
any zooming functionality on the tree. The temporary `node.inView`
property will never be updated and thus only represents the initial
state of the node when the tree was first loaded.

Therefore, any downstream uses of the inView property should be using
the `node.shell.inView` property.