Remove redux state 'geneLength'
The information for gene (CDS) lengths is better found within the
`genomeMap`, as that covers more complex genome annotations than the
`geneLength` did. Furthermore, the `geneLength` calculations included
an off-by-one bug¹ which didn't allow us to color-by the final aa in the
CDS².

The validity of genotype positions is no longer checked within color-
scale creation, due to the genomeMap not being available to that
function and a lot of prop-drilling necessary to make it so. However a
previous commit guaranteed that genotype color-bys must be valid prior
to this point, so it's ok to skip this check.

¹ In the JSON structure, both start and end are 1-based (following GFF)
so (end-start) will undercount the nuc length by 1, which resulted
in aa lengths of `<int>.6667`.

² This maps to the stop codon, so it's arguable whether this was a
feature or a bug, but there can be mutations here, and when there are
Auspice should be able to display them.