Replace geneMap with genomeMap
genomeMap better describes the annotation structure we will be using
going forward in Auspice and having one source of truth (in redux
state) is important in avoiding bugs etc. There are other pieces of
state which also contain similar information which i'll replace with
genomeMap in forthcoming commits.

There are two functional changes:

- For nucleotide positions we used to check if it fell inside a gene, which
allowed on-hover info boxes to show information such as "Nucleotide x
(Codon y in protein z)". This never considered the case where a nucleotide
fell in multiple genes (CDSs), so it was somewhat misleading.
It was a generally useful idea, so we should bring it back in an
improved form.

- CDSs which have multiple segments now have their bars (mostly)
correctly displayed -- e.g. if a CDS wraps the origin then bars will
appear at the start & end of the genome. This will be fully addressed
when a subsequent commit moves the UI to representing the CDS by itself,
rather than as segments on the genome.

Aside: the terminology used around auspice re: "gene" vs "CDS" is now
rather inconsistent. I'll hopefully tidy this up in a forthcoming
commit.