[entropy] calculate for a single CDS
When we've selected a CDS then we only want to show entropy/counts data
for that CDS (the actual UI changes will be in a forthcoming commit).
This allows us to greatly simplify `calcEntropyInView` (as well as
increasing performance).

In turn, this allows simplification of the d3 code as it is only ever
showing bars for a single CDS at a time (or the entire genome/nuc)