Guarantees about genotype colorings
We now make a guarantee at dataset-load time whether genotype colorings
are available: if mutations exist on the tree and the genomeMap exists
(which implies the `metadata.genome_annotations` exists and at least
defines the nuc coordinates) then we will always be able to have a
genotype coloring. This knowledge will allow upcoming work as we can now
be sure genotype colorings can only exist if they are supported by the
data.

Observations and bug-fixes from this commit:

- It's not possible to get a colorBy for a CDS (or 'nuc') which is not
in the genomeMap, or to specify a valid CDS but invalid position. If
such a CDS is specified in the URL query, we fallback to a default
colorBy at load time.

- It is no longer possible to have a partially-valid color-by.
Specifically, it was possible to specify multiple (genotype) positions,
and as long as one was valid then the color-by would be considered "ok".
This is now caught and the color-by changed to only those positions
which are valid, and a console.error printed.

- It's possible to have a CDS defined in the genomeMap which is never
observed on the tree. This will work as expected -- the tree coloring
will always be grey or a single color, depending on if the root-sequence
JSON is available. No entropy/counts bar will be created, so it will
appear to be zero.

- It's possible to have a CDS defined as a branch mutation which is not
present in the genomeMap, but this cannot be accessed as a coloring. It
will be shown in the branch hover/click info-box.