[entropy] Select individual CDSs
Conceptually, the model of the main axis of the entropy panel (i.e. the
axis used by the bar chart) represented the (linearised) genome. Viewing
genes/CDSs was thus a matter of zooming into that region of the genome.
This worked reasonably well for "simple" genes which comprise a single
segment but didn't work so well (or at all!) when the gene spanned
multiple segments. And for simple genes on the negative strand it was
slightly weird to zoom in but still see the translated CDS going R→L.

When viewing a CDS we now have the main axis represent this and only
this CDS, always shown going L→R. This better represents biology as we
are able to ignore any artifacts upstream of a translation; in genome
space these artifacts are often represented by multiple CDS segments
(splicing, wrapping around the origin, negative sense). Multiple
segments are still communicated via the main CDS annotation track, but
they will be back-to-back. Out-of-phase segments are handled.

The UI to view a CDS is to either click on a CDS or select a genotype
color-by for that CDS. Previous work on this branch decoupled color-by
from the selected CDS, and the intended behaviour is best seen by the
docstring of `changeEntropyCdsSelection`. In short, changing a color-by
will change the selected CDS (or show the entire genome), but changes to
the selected CDS within the entropy panel won't change the color-by. We
make some (subtle) visual changes to the annotations on the nav (lower)
axis in an attempt to indicate which CDS (segments) are selected.

The main X axis, whilst internally still using nucleotides, now shows
tick values of amino acid numbers. My desire was to use ticks which went
above and below the axis, with an additional label to communicate the
corresponding nt position in the genome. Hopefully we can implement this
in the future (the information is there in the on-hover info box of
bars).

(Most of the previous commits on this branch were specifically made to
facilitate the work here.)

NOTE: Zooming has been completely disabled in this commit. The next (?)
commit will attend to this.