[entropy] allow JSON-defined CDS segments
A JSON block, which corresponds to a CDS, may now define a segmented CDS
via a list of start/end coordinates. The order the are provided in the
JSON is crucial, as the ability for segments to wrap around the origin
prevents us from sorting them

The appropriate tests (in test/annotation-parsing.test.js) are now
passing, but -ve strand related test are still failing.