DROP! test datasets
A python script to post-process a nCoV dataset to create annotations
which more closely match the biology. There may be some off-by-one
errors there, and it's incomplete, but it's very useful for testing. Of
course we would use Nextclade to actually generate the translations

P.S. The get-data script is run for Heroku review-apps