[entropy] usability fixes
A number of small improvements, many of which were suggested in review.

Added a "help" icon, following the pattern we use extensively in the
sidebar.

Updated the zoom triangle appearance and their cursor icons. Items 1 and
4 in
<https://github.com/nextstrain/auspice/pull/1684#issuecomment-1673972542>

Changed the "return to nuc" button to a "RESET LAYOUT" button which will
always return you to a fully zoomed out view of the genome. See added
comment for why this isn't greyed out when it is a no-op. Item 2 in
<https://github.com/nextstrain/auspice/pull/1684#issuecomment-1673972542>

Changed the panel title to indicate if we're looking at the genome or a
CDS, and add the CDS name to the title in the latter case. Also added a
small axis label indicating that the numbering system is AA for selected
CDSs (it's a bit cluttered to show 'Nt #' as well)
<https://github.com/nextstrain/auspice/pull/1684#issuecomment-1679744533>