Improve data fetching error handling
The previous code would (depending on the sidecar file in question)
catch fetch errors and print console errors, and then also attempt to
load the (non-existant) data, resulting in more errors. I believe this
was due to promise chains such as:

promise.then()
  .catch()
  .then()
  .catch()

and the (mis)understanding that if the first catch clause was hit
then the entire chain would terminate.

The approach implemented here also lets us differentiate between
fetching + parsing errors, which is a subtle but nice improvement to the
warning notifications.

(Added comments are also relevant for understanding intended behaviour.)