Align first major grid circle in radial view
When plotting the first major grid, its position value can be below the
domain of the x-axis. I'm not sure why. To reduce the effects of this,
impose a lower limit on grid positions.

Previously, these situations would render a first major grid circle
offset from the center. With this change, it draws a point at the actual
center.

The text labels are similarly adjusted.