[wip] Move animation options to a collapsible section
Previously, the animation controls and options were very far away from
each other. Move the options to where the controls are and make it a
collapsible section.

Default to the section being hidden under the assumption that this is
not a commonly used feature. This allows the the default view to take up
~the same amount of vertical space as the previous layout.