tsconfig: Skip type checking all .d.ts files
This was disabled in 80835fc1dcedf051d864a8541bf7a9d7451812c6, however
there are issues with a few type declaration files in @reduxjs/toolkit
that I don't want to bother tracking down if an earlier version can be
used.

It was originally disabled with an idea that this codebase might have
some of its own declaration files, however I don't see that being
necessary ever, or at least in the near future.