Remove text summary of filters
This behaviour was restored by the previous commit but had been broken
for over 4 years. For early datasets, which used relatively few filters,
this text was mostly appropriate. For many of today's datasets it's
cumbersome, with phrases like "comprising 3 symptoms and 3 recencys".

The removal of `visibleStateCounts` should improve performance a little,
but we may wish to reinstate it in the future if we ever want to display
information about the number of visible tips per filter.