Lift node-selected modal to redux state
Lifting this from tree state to redux state paves the way to fix
inconsistent behaviour as well as allowing the information to be
displayed in a modal outside of the Tree component in the future.

We no longer update the (tip circle) radius within `clearSelectedNode`
as testing (Firefox v121, Chrome v121) shows the `onTipLeave` function
is enough to achieve the desired radius update.