Change how closing a tip modal modifies filters
Clicking on a tip activates a modal and activates the corresponding
strain filter. Previously clearing the modal would always inactivate the
filter. In the absence of any filtering this behaviour is not bad, but
when already filtering to a set of strains then it's counter-intuitive
because we inactivate a filter that was active.

The new behaviour restores the filtering state of the strain before the
modal was opened. Specifically,
* If the tip was already filtered, opening and closing the modal doesn't
  change the filters.
* If the tip was filtered but inactive, opening the modal activates it
  and closing the modal returns it to the inactive state.
* If the tip wasn't filtered (active or inactive) then opening the modal
  makes it an active filter and closing the modal removes the filter
  entirely.

Closes #1701