Restore terminal branch click behaviour
Clicking on terminal branches incorrectly behaved as if we had clicked
on the tip (the circle); the intended behaviour is for internal +
terminal branches to summarise the information the same way. This was
noticed in an issue comment
<https://github.com/nextstrain/auspice/issues/1752#issuecomment-1950527846>

The bug was introduced in f7e944dd9eee3f19c34d3611f8de916ad2465924 which
removed the ability to detect if a click occurred on a terminal node
(tip) or terminal branch.