Fix un-initialised filter console error
With the changes in the previous commit we no longer need the
`isTerminal` argument, as we use the (more accurate) `isBranch`
property. This removes the console errors "trying to remove values from
an un-initialised filter!" when shift-clicking on a branch then removing
the info modal. The bug was introduced by #1749.