Add download JSON button
The main (i.e. not sidecar) dataset is downloaded (or datasets in the
case of tangletrees). Also downloading sidecar files is also possible if
we want to expose them.

Given that we don't store the (un-modified) dataset JSON in Auspice to
download it we can either (a) convert the modified data format back to
the original JSON or (b) re-download it. (a) is by far the simpler
implementation and with appropriate caching should have good
performance.