[link out modal] refactor
The new design co-locates the data with the rendering, which is an
anti-pattern but in this case it allows for each link-out to have much
more flexibility in how it constructs the data. The co-location is
necessary for the function (component) to access `useSelector`, with the
inherent flexibility that provides. The trade-off is worth it in my
opinion.