[tangletrees] store root-sequence of second tree
In preparation for showing genotype colorings across multiple trees

We don't (yet) consider root-sequences which are defined via
sidecar files (i.e. not inline in the main JSON)

We could improve memory efficiency if we compared the nuc/aa sequences
for equality and used references to the main tree's root-sequence data,
but as we're here only considering inlined root-sequences (i.e. small
ones!) this implementation is simpler.