Update filter badges for multiple trees
The previous implementation only reflected counts in the LHS tree,
irregardless of any occurrences which may or may not be present in the
RHS tree. We now communicate both. This also fixes a bug introduced in
the previous commit where filtering to a traitName not present in the
LHS tree would cause an uncaught exception.