feat: remove missing geo_resolutions warning
When` .meta.geo_resolutions` are not present, Auspice prints a console warning:
```
JSONs did not include `geoResolutions`
```

My understanding is that it is not a required field:
https://github.com/nextstrain/augur/blob/0eab3357376f0fe7347fe60dd75dbc2fb01bed33/augur/data/schema-export-v2.json#L18
and auspice can function without it.

It seems to have been added a long time ago:
https://github.com/nextstrain/auspice/blame/a537d390e6a54214ea084240a292da786479693e/src/actions/recomputeReduxState.js#L548

I propose to remove the warning. My goal here is to clean up the dev console of Nextclade. Just being a purist here (I like my console to be clean :)), but if this warning is useful in other contexts, I don't insist.