[color-by confidence] confidence for tip colors
The previous code conveyed uncertainty in node attrs for _branches_ by
making them appear grey-er, but we never implemented this for _tips_;
most likely because we never had a dataset with such data when this was
built.

Here we use the same approach for tips as for branches, but with a
slightly different parameterisation of the interpolation. The mapping of
the entropy value into `[0,1]` (`tipOpacityFunction`) was chosen so that
tips with no (or very little) uncertainty look unchanged from previous
Auspice versions, and uncertainty makes them appear more similar to the
branch colour (for an equivalent uncertainty).