Tip label can be any node_attr
Any normal `node_attr` in the tree can be used as a tip label, as well
as the special-cases of strain-name and "none". Previously we only
allowed valid colorings. Normal here means a node_attr value which is an
object with a "value" key, thus excluding special-cases such as "div",
"hidden", "vaccine" etc - see the JSON schema for more context
<https://github.com/nextstrain/augur/blob/master/augur/data/schema-export-v2.json>