measurements: Add URL query param `mf_<field>`
Use query param `mf_<field>=<value>` to specify active filters for
the measurements panel. Invalid fields or values are ignored and
removed.

Multiple values for the same field are expected to give the query param
multiple times with different values, e.g.
`mf_reference_strain=A/Stockholm/18/2011&mf_reference_strain=A/Alabama/5/2010`
This is slightly different than the behavior of the tree filter param,
but I'm following this pattern to avoid the issue described in
<https://github.com/nextstrain/auspice/issues/1846>