measurements: add URL query param `m_overallMean`
Lays the foundation for measurements URL query params, starting with the
simplest `m_overallMean` to toggle the display of the overall mean.

A series of decisions made in this commit:

1. Keeps conversion between query params and control states in the same
file so that they are less likely to be out of sync.
2. Uses a single dispatch action to create both the new control state
and query params to make sure they are in-sync.
3. If the new control state matches the defaults, then the query param
is removed. This matches the behavior of other query params in Auspice.