measurements: Add URL query param `m_collection`
Use query param `m_collection=<collection_key>` to specify which
collection to display in the measurements panel. The display will use
the collection's display defaults unless there are other measurements
query params included in the URL.

If the collection key is the default collection, then the query param
will be removed. If the collection key is invalid, then the query param
will be removed and the default collection will be displayed.