Add types to narrative state
This was necessary to properly type the root state.

Note: AnyAction is used for `action` which is only slightly better than
any. Ideally, each action type would come with its own TypeScript
interface but there is no good way to tightly couple those. I think the
right improvement here is to use Redux's createSlice but that is beyond
the scope of adding types.