[entropy] improve performance of entropy calcs
by not recalculating data on the leading edge of the debounced
recalculation action. For big trees (which take >500ms of time to
redraw) the main thread is still blocked for roughly the same amount of
time, but the tree is redrawn faster. For small trees which redraw
quicker than that the entropy doesn't update until the debounce 500ms
timeout is reached, resulting in slightly odd behaviour.

(Reducing this timout also results in less-than-ideal behaviour as
(e.g.) dragging the date range of a tree results in interruptions while
the entropy calcs run which is worse IMO.)