[entropy] skip recalc if panel's not rendered
Improves the performance of most interactions in Auspice when the
`<Entropy>` panel's not displayed by not constantly updating the entropy
calculations. The downside is when the panel's toggled on things are a
little slower as we recalculate at that point.

Note that the Entropy panel should never be rendered with invalid /
stale / uncomputed data. Any action which results in the panel being
shown is responsible for ensuring the underlying data is updated as
appropriate.